requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.3.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
